Refactoring the program to work better.
