+++ Cauldron.py
@@ -1,9 +1,6 @@
-from typing import Callable
 import json
 import os
 
-def clear():
-    os.system("cls")
 
 class FileManagment:
     current_brew = None
@@ -32,45 +29,24 @@ class FileManagment:
         with open("Brews\\"+brew["_meta"]["sources"][0]["json"]+".json", "w") as file:
                 json.dump(brew, file, sort_keys=True, indent=4)
 
-class Navigation: 
-    def menu_generator(options: list[tuple[str, Callable]]) -> Callable:
-        '''Generates a Menu from a list of tuples.
 
-        Args:
-            options (list[tuple[str, Callable]]): str is for the name of the option and callable is a function that will be returned if the item is selected.
-
-        Returns:
-            Callable: Returns the selected function.
-        '''
-        while True:
-            clear()
-            for index, (option, _) in enumerate(options):
-                print(f"({index + 1}) - {option}")
-            try:
-                choice = int(input("Select an option: "))
-                if 1 <= choice <= len(options):
-                    return options[choice - 1][1]
-                else:
-                    print("Invalid choice. Please try again.")
-            except ValueError:
-                print("Invalid input. Please enter a number.")
   
-    def new_meta_setup(brew):
-        clear()
-        sources = brew["_meta"]["sources"][0]
-        input("The setup process for the basic file data is now taking place. This will be documentation related to your homebrew.\nIf you make a mistake or change your mind you will be able to edit this later.\nPress ENTER to continue.")
-        clear()
-        sources["json"] = input("Please enter a json identifier for your brew file. It must be completely unique to any other homebrew.\nThis is only used as an identifier and will not show up anywhere other than here. Just make it related and unique to the project.\n(Minimum of 6 characters, No Spaces, No Symbols, No Numbers)\nJSON: ")
-        clear()
-        sources["abbreviation"] = input("Please enter an abbreviation for your brew. (e.g. PHB for 'Player Hand Book')\nAbbreviation: ")
-        clear()
-        sources["full"] = input("Please enter the full title of the Brew.\nBrew Name: ")
-        clear()
-        sources["authors"] = input("Please enter the author names for contributers to the source material.\n(Format is as follows: Bob Bowen, Carl Coolguy)\nAuthors: ").split(", ")
-        clear()
-        sources["convertedBy"] = input("Please enter the names of who converted the source material into a 5eTools compatible json file. (This will most likely just be you since your using this program. Feel free to credit the creator of the program though.)\n(Format is as follows: Bob Bowen, Carl Coolguy)\nConverters: ").split(", ")
-        brew["_meta"]["sources"][0] = sources
-        return brew
+def new_meta_setup(brew):
+    clear()
+    sources = brew["_meta"]["sources"][0]
+    input("The setup process for the basic file data is now taking place. This will be documentation related to your homebrew.\nIf you make a mistake or change your mind you will be able to edit this later.\nPress ENTER to continue.")
+    clear()
+    sources["json"] = input("Please enter a json identifier for your brew file. It must be completely unique to any other homebrew.\nThis is only used as an identifier and will not show up anywhere other than here. Just make it related and unique to the project.\n(Minimum of 6 characters, No Spaces, No Symbols, No Numbers)\nJSON: ")
+    clear()
+    sources["abbreviation"] = input("Please enter an abbreviation for your brew. (e.g. PHB for 'Player Hand Book')\nAbbreviation: ")
+    clear()
+    sources["full"] = input("Please enter the full title of the Brew.\nBrew Name: ")
+    clear()
+    sources["authors"] = input("Please enter the author names for contributers to the source material.\n(Format is as follows: Bob Bowen, Carl Coolguy)\nAuthors: ").split(", ")
+    clear()
+    sources["convertedBy"] = input("Please enter the names of who converted the source material into a 5eTools compatible json file. (This will most likely just be you since your using this program. Feel free to credit the creator of the program though.)\n(Format is as follows: Bob Bowen, Carl Coolguy)\nConverters: ").split(", ")
+    brew["_meta"]["sources"][0] = sources
+    return brew
         
 class Editor:
     def edit_schema():
