+++ filemanagment.py
@@ -1,0 +1,8 @@
+"""
+This module is for all things related to all things json file.
+"""
+
+def load_file(a: str, path="") -> dict:
+    file = None
+    if path:
+
